requirements.txt
pulp>=2.8.0
numpy>=1.26.0
pytest>=8.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
